* Fix state header. This should belong to an earlier commit.

+++ redshift/_state.pxd
@@ -46,7 +46,7 @@ cdef struct SlotTokens:
     size_t pscopy
     size_t psexact
 
-    size_t n0_prob
+    int n0_prob
 
 
 cdef struct State:
